Relocate _load and _dump (#2)
