+++ src/cleantest/pkg/charmlib.py
@@ -6,16 +6,10 @@
 
 from __future__ import annotations
 
-import hashlib
 import json
-import os
-import pathlib
-import pickle
 import subprocess
-import tempfile
-import uuid
 from shutil import which
-from typing import Dict, List
+from typing import List
 
 from cleantest._utils import detect_os_variant
 from cleantest.pkg._base import Package, PackageError
@@ -53,19 +47,6 @@ class Charmlib(Package):
             self._auth_token = _manager._auth_token
             self._charmlib_store = _manager._charmlib_store
 
-    @classmethod
-    def _load(cls, _manager: str, hash: str) -> object:
-        if type(_manager) == str and os.path.isfile(_manager):
-            fin = pathlib.Path(_manager)
-            if hash != hashlib.sha224(fin.read_bytes()).hexdigest():
-                raise PackageError("SHA224 hashes do not match. Will not load untrusted object.")
-
-            return cls(_manager=pickle.loads(fin.read_bytes()))
-        else:
-            raise PackageError(
-                f"Invalid type {type(_manager)} received. Type must either be str or bytes."
-            )
-
     def _run(self) -> None:
         self._setup()
         self.__handle_charm_lib_install()
@@ -100,15 +81,6 @@ class Charmlib(Package):
                 raise PackageError(
                     f"Failed to install charmcraft using the following command: {' '.join(cmd)}"
                 )
-
-    def _dump(self) -> Dict[str, str]:
-        """Return a path to a pickled object and hash for verification."""
-        filepath = os.path.join(tempfile.gettempdir(), f"{uuid.uuid4()}.pkl")
-        data = pickle.dumps(self)
-        hash = hashlib.sha224(data).hexdigest()
-        fout = pathlib.Path(filepath)
-        fout.write_bytes(pickle.dumps(self))
-        return {"path": filepath, "hash": hash}
 
     def __handle_charm_lib_install(self) -> None:
         env = {"CHARMCRAFT_AUTH": self._auth_token}
